Fixed DHT, added ADC+CO sensor.

+++ project/src/Main/COSensor.hpp
@@ -4,7 +4,7 @@
 namespace COSensor
 {
     bool initializeSensor();
-    int16_t getCarbonMonoxidePartsPerMillion();
+    uint32_t getCarbonMonoxidePartsPerMillion();
 }
 
 #endif // CO_SENSOR_HPP
